change method name

+++ library_generation/test/utilities_unit_tests.py
@@ -313,7 +313,7 @@ class UtilitiesTest(unittest.TestCase):
         file_comparator.compare_files(
             f"{library_path}/owlbot.py", f"{library_path}/owlbot-golden.py"
         )
-        self.__remove_generated_files(path=library_path, is_monorepo=False)
+        self.__remove_prerequisite_files(path=library_path, is_monorepo=False)
 
     def test_generate_prerequisite_files_monorepo_success(self):
         library_path = self.__setup_prerequisite_files(num_libraries=2)
@@ -329,7 +329,7 @@ class UtilitiesTest(unittest.TestCase):
         file_comparator.compare_files(
             f"{library_path}/owlbot.py", f"{library_path}/owlbot-golden.py"
         )
-        self.__remove_generated_files(path=library_path)
+        self.__remove_prerequisite_files(path=library_path)
 
     def test_prepare_repo_monorepo_success(self):
         gen_config = self.__get_a_gen_config(2)
@@ -375,7 +375,7 @@ class UtilitiesTest(unittest.TestCase):
         self.assertEqual("output", Path(repo_config.output_folder).name)
         library_path = sorted([Path(key).name for key in repo_config.libraries])
         self.assertEqual(["misc"], library_path)
-        shutil.rmtree(repo_config.output_folder, ignore_errors=True)
+        shutil.rmtree(repo_config.output_folder)
 
     def __setup_prerequisite_files(
         self, num_libraries: int, library_type: str = "GAPIC_AUTO"
@@ -452,13 +452,14 @@ class UtilitiesTest(unittest.TestCase):
         )
 
     @staticmethod
-    def __remove_generated_files(path: str, is_monorepo: bool = True) -> None:
+    def __remove_prerequisite_files(path: str, is_monorepo: bool = True) -> None:
         os.remove(f"{path}/.repo-metadata.json")
         os.remove(f"{path}/owlbot.py")
         if is_monorepo:
             os.remove(f"{path}/.OwlBot-hermetic.yaml")
             return
-        shutil.rmtree(f"{path}/.github", ignore_errors=True)
+        if os.path.isdir(f"{path}/.github"):
+            shutil.rmtree(f"{path}/.github", ignore_errors=True)
 
 
 if __name__ == "__main__":
